services/harborsim/requirements/deweaponize.txt
beautifulsoup4==4.12.3
bleach==6.1.0
lxml==5.3.0
boto3==1.35.0
pyyaml==6.0.2

